+++ src/main/java/com/cmeza/deployer/plugin/copy/CopyTarget.java
@@ -1,24 +1,17 @@
 package com.cmeza.deployer.plugin.copy;
 
+import com.cmeza.deployer.plugin.copy.configurations.CopyBundle;
+import com.cmeza.deployer.plugin.utils.AbstractTarget;
 import lombok.Data;
 import org.apache.maven.plugins.annotations.Parameter;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Data
-public class CopyTarget {
+public class CopyTarget extends AbstractTarget<CopyBundle> {
 
-    @Parameter(name = "destinationFolder")
-    private String destinationFolder;
-
-    @Parameter(name = "searchIn")
-    private String searchIn;
-
-    @Parameter(name = "findInParent", defaultValue = "false")
-    private boolean findInParent;
-
-    @Parameter(name = "prefix")
-    private String prefix;
-
-    @Parameter(name = "suffix")
-    private String suffix;
+    @Parameter(name = "bundles")
+    private List<CopyBundle> bundles = new ArrayList<>();
 
 }