+++ src/main/java/com/cmeza/deployer/plugin/exec/constants/ExecConstants.java
@@ -1,0 +1,5 @@
+package com.cmeza.deployer.plugin.exec.constants;
+
+public class ExecConstants {
+    public static final boolean VERBOSE = false;
+}