+++ src/main/java/com/cmeza/deployer/plugin/utils/IAbstractTarget.java
@@ -1,0 +1,9 @@
+package com.cmeza.deployer.plugin.utils;
+
+import java.util.List;
+
+public interface IAbstractTarget<T extends AbstractBundle> {
+    String getBundleConfiguration();
+
+    List<T> getBundles();
+}