+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Change Log
 
+## [1.1.0] - 2021-12-31
+### Added
+- Exec command (Windows/Linux)
+- Configuration exec by json/xml
+- Configuration copy by json/xml
+- Configuration minify by json/xml
+
 ## [1.0.0] - 2021-12-28
 ### Added
 - Copy files/folders and minify(css, js)
 
+[1.1.0]: https://github.com/cmeza20/copy-minify-maven-plugin/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/cmeza20/copy-minify-maven-plugin/tree/1.0.0